+++ CHANGELOG.md
@@ -7,8 +7,9 @@ Version 1.0.0
 
 1. Initial Release of repository
 2. Formula Changes
-    1. None
+    1. Added temporary saml2aws Formula for JumpCloud/Duo support
 3. Cask Changes
     1. CAmod 0.61 release
+    2. openra-playtest added
 
 Return to [README](README.md)